+++ packages/snjs/lib/application_interfaces/listed_interface.ts
@@ -1,16 +1,5 @@
+import { ListedAccount, ListedAccountInfo } from '@Lib/services/api/responses';
 import { UuidString } from '@Lib/types';
-
-export type ListedAccount = {
-  secret: string;
-  authorId: string;
-  hostUrl?: string;
-};
-
-export type ListedAccountInfo = {
-  display_name: string;
-  author_url: string;
-  settings_url: string;
-};
 
 export interface ListedInterface {
   canRegisterNewListedAccount: () => boolean;
