+++ packages/snjs/lib/models/index.ts
@@ -12,7 +12,7 @@ export {
   ActionsExtensionMutator,
   SNActionsExtension,
 } from '@Models/app/extension';
-export { Action } from '@Models/app/action';
+export type { Action } from '@Models/app/extension';
 export { SNNote, NoteMutator } from '@Models/app/note';
 export { SNTag, TagMutator } from '@Models/app/tag';
 export { SNUserPrefs, UserPrefsMutator, PrefKey } from '@Models/app/userPrefs';
@@ -20,7 +20,5 @@ export type { PrefValue } from '@Models/app/userPrefs';
 export { SNSmartTag } from '@Models/app/smartTag';
 export { SNTheme, ThemeMutator } from '@Models/app/theme';
 
-export {
-  displayStringForContentType,
-} from '@Models/content_types';
+export { displayStringForContentType } from '@Models/content_types';
 export { CreateItemFromPayload } from '@Models/generator';
